Added Alerts and Fixed Some Spelling Errors
It should now be possible to add Alerts to the database. These are
associated with a doctor it should be sent to and a condition that
caused the alert. updateAppoinemt is now updateAppointment as it
should be. I have also removed the requirement of giving an old
version to update methods as the id should never change.